Fixes Context Keys
This patch updates the tests/benchmarks/examples/README all to use the
context key functions instead of the constants directly. This ensures
the context keys are already on the heap and are not placed on the heap
every time they are wrapped inside the interface{} type when invoking
context.WithValue or ctx.Value.